Fix conditional compilation errors and split up reified propagator
