+++ pumpkin-crates/core/src/api/mod.rs
@@ -69,8 +69,8 @@ pub mod options {
     pub use crate::engine::ConflictResolverType;
     pub use crate::engine::RestartOptions;
     pub use crate::engine::SatisfactionSolverOptions as SolverOptions;
+    pub use crate::propagators::ReifiedPropagatorArgs;
     pub use crate::propagators::nogoods::LearningOptions;
-    pub use crate::propagators::reified_propagator::ReifiedPropagatorArgs;
 }
 
 pub mod termination {
